Version 1.7.1 Add CachedPostStats model, add cron job for insert cache query results.

+++ src/main/java/com/indiepost/repository/StatMetadataRepository.java
@@ -1,10 +1,11 @@
 package com.indiepost.repository;
 
-import com.indiepost.model.StatMetadata;
+import com.indiepost.dto.StatMetadata;
 import org.springframework.data.repository.CrudRepository;
 
 /**
- * Created by jake on 10/24/17.
+ * Created by jake on 10/29/17.
  */
-public interface StatMetadataRepository extends CrudRepository<StatMetadata, Integer> {
+public interface StatMetadataRepository extends CrudRepository<StatMetadata, Long> {
+
 }
